pyproject.toml
[tool.poetry]
name = "python-project-initial"
version = "0.0.1"
description = "python项目脚手架"
authors = ["WalterZhu <hqzhu0461@163.com>"]
readme = "README.md"
packages = [
    { include = "*", from = "src"},
    { include = "main.py",  from = ""}
]

[tool.poetry.scripts]
server = "main:main"

[tool.poetry.dependencies]
python = "^3.11"
pytest = "^7.2.1"
python-dotenv = "^1.0.0"

[tool.pytest.ini_options]
addopts = "-v -W ignore::DeprecationWarning"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
